# src/main/resources/docs/dgbank-api-docs.yaml
openapi: 3.0.3
info:
  title: DG Bank API
  description: |
    API completa per il sistema bancario DG Bank.
    
    ## Autenticazione
    L'API utilizza JWT Bearer token per l'autenticazione. Dopo il login o la registrazione, includi il token nell'header Authorization:
    `Authorization: Bearer <token>`
    
    ## Rate Limiting
    - Token JWT: 24 ore di validità
    - Refresh token: 7 giorni di validità
    
    ## Formati Monetari
    Tutti gli importi sono gestiti con precisione decimale (15,2) per evitare errori di arrotondamento.
  version: 1.0.0
  contact:
    name: DG Bank Development Team
    email: dev@dgbank.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.dgbank.com
    description: Production server
...
paths:
...
  /cards/issue:
    post:
      tags:
        - Cards
      summary: Emetti nuova carta
      description: |
        Emette una nuova carta collegata ad un conto esistente.
        Genera automaticamente numero carta, CVV e data di scadenza (4 anni).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewCardRequest'
      responses:
        '201':
          description: Carta emessa con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '400':
          description: Errore nell'emissione della carta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Account non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /cards/{cardId}:
    get:
      tags:
        - Cards
      summary: Dettagli carta
      description: Ottieni i dettagli di una carta specifica (numero mascherato per sicurezza)
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Dettagli della carta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '404':
          description: Carta non trovata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /cards/user/{userId}:
    get:
      tags:
        - Cards
      summary: Carte utente
      description: Ottieni tutte le carte di un utente
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Lista delle carte dell'utente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  cards:
                    type: array
                    items:
                      $ref: '#/components/schemas/Card'
  /cards/account/{accountId}:
    get:
      tags:
        - Cards
      summary: Carte del conto
      description: Ottieni tutte le carte collegate ad un conto specifico
      security:
        - bearerAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Lista delle carte del conto
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  cards:
                    type: array
                    items:
                      $ref: '#/components/schemas/Card'
  /cards/{cardId}/block:
    put:
      tags:
        - Cards
      summary: Blocca carta
      description: Blocca una carta (imposta isActive = false)
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Carta bloccata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Errore nel blocco della carta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /cards/{cardId}/activate:
    put:
      tags:
        - Cards
      summary: Attiva carta
      description: Attiva una carta (imposta isActive = true)
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Carta attivata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Errore nell'attivazione della carta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Transaction endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT ottenuto tramite login o registrazione.
        Durata: 24 ore per access token, 7 giorni per refresh token.
  schemas:
...
    NewCardRequest:
      type: object
      required:
        - accountId
      properties:
        accountId:
          type: integer
          format: int64
          description: ID del conto a cui collegare la carta
...
    CardResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Carta creata con successo"
        card:
          $ref: '#/components/schemas/Card'
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operazione completata con successo"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Messaggio di errore"

    # Entity models
...
    Card:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID univoco della carta
        cardNumber:
          type: string
          description: Numero carta mascherato per sicurezza
          example: "**** **** **** 3456"
        expiryDate:
          type: string
          format: date
          description: Data di scadenza della carta
        isActive:
          type: boolean
          description: Stato della carta (true = attiva, false = bloccata)
        owner:
          type: string
          description: Nome del proprietario della carta
          example: "Mario Rossi"
        accountNumber:
          type: string
          description: Numero IBAN del conto collegato
          example: "IT 60 X05472 81110 000000123456"
...
security:
  - bearerAuth: []

# Esempi di utilizzo
...